// src/screens/deck-form/deck-form/store/card-form-store-interface.ts
import { CardFormType } from "./deck-form-store.ts";
import { TextField } from "mobx-form-lite";
import { DeckSpeakFieldEnum } from "../../../../../functions/db/deck/decks-with-cards-schema.ts";

export type CardInnerScreenType =
  | "cardPreview"
  | "cardType"
  | "example"
  | "aiSpeech"
  | null;

export type LimitedDeckForm = {
  speakingCardsLocale: TextField<string | null>;
  speakingCardsField: TextField<DeckSpeakFieldEnum | null>;
  cardInputModeId: string | null;
};

export interface CardFormStoreInterface {
  cardForm?: CardFormType | null;
  onSaveCard: () => void;
  onBackCard: () => void;
  cardInnerScreen: TextField<CardInnerScreenType>;
  isSending: boolean;
  markCardAsRemoved?: () => void;

  deckForm?: LimitedDeckForm;

  // Navigation next and previous card
  isPreviousCardVisible?: boolean;
  isNextCardVisible?: boolean;
  onPreviousCard?: () => void;
  onNextCard?: () => void;
}